refactor(tests): split auth middleware integration suite for GH727
